Move sleep to src
