+++ .github/shared/package.json
@@ -11,8 +11,8 @@
     "./logger": "./src/logger.js",
     "./readme": "./src/readme.js",
     "./sdk-types": "./src/sdk-types.js",
-    "./test/examples": "./test/examples.js",
-    "./test/sleep": "./test/sleep.js"
+    "./sleep": "./src/sleep.js",
+    "./test/examples": "./test/examples.js"
   },
   "_comments": {
     "dependencies": "Runtime dependencies must be kept to an absolute minimum for performance, ideally with no transitive dependencies",
