+++ .github/shared/test/sleep.test.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { describe, expect, it } from "vitest";
-import { sleep } from "./sleep.js";
+import { sleep } from "../src/sleep.js";
 
 describe("sleep", () => {
   it.each([-1, 0, 1, 2])("sleep(%s)", async (ms) => {
